requirements.txt
django>=4.2,<5.2
wagtail>=6.4,<6.5
dj-database-url
psycopg[binary]
whitenoise
gunicorn==23.0.0
django-storages[s3]
wagtail-storages
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-environ>=0.10.0
